+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Additional documentation and release notes are available at [Multiplayer Documen
 ## [Unreleased] - yyyy-mm-dd
 
 ### Added
-*
+* Dedicated Server Mode (#696) - This PR adds the new flows necessary for boss room to start a dedicated server. Please read PR for more details.
+* 
 ### Changed
 * Updated tools, authentication and relay packages (#690)
 * Replaced our dependency injection solution with VContainer. (#679)